requirements 2..txt
Adafruit-Blinka                          8.56.0
adafruit-circuitpython-busdevice         5.2.11
adafruit-circuitpython-connectionmanager 3.1.3
adafruit-circuitpython-dht               4.0.7
adafruit-circuitpython-framebuf          1.6.7
adafruit-circuitpython-requests          4.1.10
adafruit-circuitpython-ssd1306           2.12.19
adafruit-circuitpython-typing            1.11.2
Adafruit-PlatformDetect                  3.77.0
Adafruit-PureIO                          1.1.11
aiohappyeyeballs                         2.6.1
aiohttp                                  3.11.16
aiohttp-retry                            2.9.1
aiosignal                                1.3.2
altair                                   5.5.0
attrs                                    25.3.0
binho-host-adapter                       0.1.6
bitarray                                 3.3.1
bitstring                                4.3.1
blinker                                  1.9.0
cachetools                               5.5.2
camera                                   1.3.0
certifi                                  2025.1.31
cffi                                     1.17.1
charset-normalizer                       3.4.1
click                                    8.1.8
colorama                                 0.4.6
contourpy                                1.3.1
cryptography                             44.0.2
cycler                                   0.12.1
dnspython                                2.7.0
ecdsa                                    0.19.1
esptool                                  4.8.1
filelock                                 3.18.0
Flask                                    3.1.0
fonttools                                4.57.0
frozenlist                               1.5.0
fsspec                                   2025.3.2
gitdb                                    4.0.12
GitPython                                3.1.44
idna                                     3.10
intelhex                                 2.3.0
itsdangerous                             2.2.0
Jinja2                                   3.1.5
joblib                                   1.4.2
jsonschema                               4.23.0
jsonschema-specifications                2024.10.1
kiwisolver                               1.4.8
MarkupSafe                               3.0.2
matplotlib                               3.10.1
mpmath                                   1.3.0
multidict                                6.4.3
narwhals                                 1.34.0
network                                  0.1
networkx                                 3.4.2
numpy                                    2.1.1
opencv-contrib-python                    4.11.0.86
opencv-python                            4.11.0.86
packaging                                24.2
pandas                                   2.2.3
pillow                                   11.1.0
pip                                      25.0.1
propcache                                0.3.1
protobuf                                 5.29.4
psutil                                   7.0.0
py-cpuinfo                               9.0.0
pyarrow                                  19.0.1
pycparser                                2.22
pydeck                                   0.9.1
pyftdi                                   0.56.0
PyJWT                                    2.10.1
pymongo                                  4.11.1
pyparsing                                3.2.3
pyserial                                 3.5
python-dateutil                          2.9.0.post0
pytz                                     2025.2
pyusb                                    1.3.1
PyYAML                                   6.0.2
reedsolo                                 1.7.0
referencing                              0.36.2
requests                                 2.32.3
rpds-py                                  0.24.0
scikit-learn                             1.6.1
scipy                                    1.15.2
seaborn                                  0.13.2
setuptools                               78.1.0
six                                      1.17.0
smmap                                    5.0.2
streamlit                                1.44.1
streamlit-autorefresh                    1.0.1
streamlit-lottie                         0.0.5
sympy                                    1.13.1
tenacity                                 9.1.2
thop                                     0.1.1
threadpoolctl                            3.6.0
toml                                     0.10.2
torch                                    2.6.0
torchvision                              0.21.0
tornado                                  6.4.2
tqdm                                     4.67.1
twilio                                   9.5.2
typing_extensions                        4.13.1
tzdata                                   2025.2
ultralytics                              8.3.107
ultralytics-thop                         2.0.14
urllib3                                  2.3.0
watchdog                                 6.0.0
Werkzeug                                 3.1.3
yarl                                     1.19.0
